+++ components/compositing/Cargo.toml
@@ -41,7 +41,6 @@ stylo_traits = { workspace = true }
 tracing = { workspace = true, optional = true }
 webrender = { workspace = true }
 webrender_api = { workspace = true }
-webrender_traits = { workspace = true }
 webxr = { path = "../webxr", optional = true }
 
 [dev-dependencies]
