+++ components/layout_2020/display_list/mod.rs
@@ -8,6 +8,7 @@ use std::sync::Arc;
 use app_units::Au;
 use base::WebRenderEpochToU16;
 use base::id::ScrollTreeNodeId;
+use compositing_traits::display_list::{AxesScrollSensitivity, CompositorDisplayListInfo};
 use embedder_traits::Cursor;
 use euclid::{Point2D, SideOffsets2D, Size2D, UnknownUnit};
 use fonts::GlyphStore;
@@ -34,7 +35,6 @@ use webrender_api::{
     self as wr, BorderDetails, BoxShadowClipMode, ClipChainId, CommonItemProperties,
     ImageRendering, NinePatchBorder, NinePatchBorderSource, units,
 };
-use webrender_traits::display_list::{AxesScrollSensitivity, CompositorDisplayListInfo};
 use wr::units::LayoutVector2D;
 
 use crate::context::{LayoutContext, ResolvedImage};
