+++ components/media/lib.rs
@@ -9,16 +9,16 @@ mod media_thread;
 
 use std::sync::{Arc, Mutex};
 
+use compositing_traits::{
+    WebrenderExternalImageApi, WebrenderExternalImageHandlers, WebrenderExternalImageRegistry,
+    WebrenderImageHandlerType, WebrenderImageSource,
+};
 use euclid::default::Size2D;
 use ipc_channel::ipc::{IpcReceiver, IpcSender, channel};
 use log::warn;
 use serde::{Deserialize, Serialize};
 use servo_config::pref;
 pub use servo_media::player::context::{GlApi, GlContext, NativeDisplay, PlayerGLContext};
-use webrender_traits::{
-    WebrenderExternalImageApi, WebrenderExternalImageHandlers, WebrenderExternalImageRegistry,
-    WebrenderImageHandlerType, WebrenderImageSource,
-};
 
 use crate::media_thread::GLPlayerThread;
