+++ components/script/Cargo.toml
@@ -49,6 +49,7 @@ canvas_traits = { workspace = true }
 cbc = { workspace = true }
 chrono = { workspace = true }
 cipher = { workspace = true }
+compositing_traits = { workspace = true }
 constellation_traits = { workspace = true }
 content-security-policy = { workspace = true }
 cookie = { workspace = true }
@@ -134,7 +135,6 @@ uuid = { workspace = true, features = ["serde"] }
 webdriver = { workspace = true }
 webgpu_traits = { workspace = true }
 webrender_api = { workspace = true }
-webrender_traits = { workspace = true }
 webxr-api = { workspace = true, features = ["ipc"], optional = true }
 wgpu-core = { workspace = true }
 wgpu-types = { workspace = true }
