+++ components/shared/compositing/rendering_context.rs
