+++ components/shared/script_layout/Cargo.toml
@@ -16,6 +16,7 @@ base = { workspace = true }
 app_units = { workspace = true }
 atomic_refcell = { workspace = true }
 canvas_traits = { workspace = true }
+compositing_traits = { workspace = true }
 constellation_traits = { workspace = true }
 embedder_traits = { workspace = true }
 euclid = { workspace = true }
@@ -41,4 +42,3 @@ servo_url = { path = "../../url" }
 stylo = { workspace = true }
 stylo_traits = { workspace = true }
 webrender_api = { workspace = true }
-webrender_traits = { workspace = true }
