--- components/shared/webrender/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "webrender_traits"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-publish.workspace = true
-rust-version.workspace = true
-
-[lib]
-name = "webrender_traits"
-path = "lib.rs"
-test = true
-
-[features]
-no-wgl = ["surfman/sm-angle-default"]
-
-[dependencies]
-base = { workspace = true }
-dpi = { version = "0.1" }
-embedder_traits = { workspace = true }
-euclid = { workspace = true }
-gleam = { workspace = true }
-glow = { workspace = true }
-image = { workspace = true }
-ipc-channel = { workspace = true }
-libc = { workspace = true }
-log = { workspace = true }
-raw-window-handle = { version = "0.6" }
-serde = { workspace = true }
-servo-media = { workspace = true }
-servo_geometry = { path = "../../geometry" }
-stylo = { workspace = true }
-surfman = { workspace = true, features = ["sm-x11"] }
-webrender_api = { workspace = true }
