+++ components/webgpu/Cargo.toml
@@ -14,6 +14,7 @@ path = "lib.rs"
 [dependencies]
 arrayvec = { workspace = true, features = ["serde"] }
 base = { workspace = true }
+compositing_traits = { workspace = true }
 euclid = { workspace = true }
 ipc-channel = { workspace = true }
 log = { workspace = true }
@@ -23,7 +24,6 @@ servo_config = { path = "../config" }
 webgpu_traits = { workspace = true }
 webrender = { workspace = true }
 webrender_api = { workspace = true }
-webrender_traits = { workspace = true }
 wgpu-core = { workspace = true, features = ["serde", "wgsl"] }
 wgpu-types = { workspace = true }
